requirements-dev.txt
nox==2022.1.7
pre-commit==2.15.0
pydantic-to-typescript==1.0.8
datamodel-code-generator==0.13.1
